+++ cmd/sycontainerize/sycontainerize.go
@@ -15,11 +15,11 @@ import (
 
 	"github.com/gvallee/go_util/pkg/util"
 	"github.com/gvallee/kv/pkg/kv"
-	"github.com/sylabs/singularity-mpi/internal/pkg/checker"
-	"github.com/sylabs/singularity-mpi/internal/pkg/launcher"
-	"github.com/sylabs/singularity-mpi/internal/pkg/sy"
-	"github.com/sylabs/singularity-mpi/internal/pkg/sys"
+	"github.com/sylabs/singularity-mpi/pkg/checker"
 	"github.com/sylabs/singularity-mpi/pkg/containerizer"
+	"github.com/sylabs/singularity-mpi/pkg/launcher"
+	"github.com/sylabs/singularity-mpi/pkg/sy"
+	"github.com/sylabs/singularity-mpi/pkg/sys"
 )
 
 func main() {
