+++ pkg/checker/distro.go
